tests/wpt/tests/import-maps/static-after-innerText.html

<!DOCTYPE html>
<html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script type="importmap" id="importMap">
{
  "imports": {
    "./resources/log.js?pipe=sub&name=A": "./resources/log.js?pipe=sub&name=B"
  }
}
</script>
<script>
const log = [];

promise_test(() => {
  // Import maps are static. Changing their innerText does not change the import
  // map entry.
  const importMap = document.getElementById("importMap");
  importMap.innerText = "";

  // The following should succeed as if the above lines did nothing.
  return import("./resources/log.js?pipe=sub&name=A")
    .then(() => import("./resources/log.js?pipe=sub&name=B"))
    .then(() => assert_array_equals(log, ["log:B"]))
  },
  "Module map's key is the URL after import map resolution");
</script>
